expiryguard-backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.7.0
pydantic[email]==2.11.1
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
APScheduler==3.11.0
httpx==0.28.1
reportlab==4.2.5
pytesseract==0.3.13
Pillow==11.1.0
python-dotenv==1.1.0
pymongo==4.11.3
